+++ src/types/convert.rs
@@ -62,7 +62,7 @@ impl TryFrom<UpdateScene> for scene::SceneSpec {
     // very temporary and inefficient
     fn try_from(update_scene: UpdateScene) -> Result<Self, Self::Error> {
         fn try_from_node(node: Component) -> Result<Vec<scene::NodeSpec>, TypeError> {
-            if let node::ComponentParams::InputStream(_) = node.params {
+            if let component::ComponentParams::InputStream(_) = node.params {
                 return Ok(vec![]);
             }
             let spec = node.clone().try_into()?;
